Development

+++ Backend/WebshopApi/Controllers/WishlistController.cs
@@ -1,5 +1,9 @@
+
+using System.Security.Claims;
+
 using DataTransferObject;
 
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 using models;
@@ -22,35 +26,67 @@ public class WishlistController : ControllerBase
     [HttpGet]
     public async Task<ActionResult<List<Wishlists>>> GetAllWishlists()
     {
-        var wishlists = _wishlistservice.GetAllWishLists();
+        var wishlists = await _wishlistservice.GetAllWishLists();
         return Ok(wishlists);
     }
 
     [HttpGet("{id:int}")]
-    public async Task<ActionResult<Wishlists>> GetWishlistById(int id)
+    public async Task<ActionResult<List<Wishlists>>> GetWishlistById(int id)
     {
-        var wishlist = _wishlistservice.GetWishlistsById(id);
+        var wishlist = await _wishlistservice.GetWishlistsById(id);
         return Ok(wishlist);
     }
-
+    [Authorize]
+    [HttpGet("mine")]
+    public async Task<ActionResult<Wishlists>> GetMyWishlist()
+    {
+        var useridstring = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        // return Ok(useridstring);
+        if (string.IsNullOrEmpty(useridstring)) return Unauthorized();
+        var userId = int.Parse(useridstring);
+        var wishlists = await _wishlistservice.GetWishlistsByUserId(userId);
+        return Ok(wishlists);
+    }
+    [Authorize]
     [HttpPost("create")]
     public async Task<ActionResult> CreateWishlist(WishlistDTO wishlistDTO)
     {
-        _wishlistservice.CreateService(wishlistDTO);
-        return Ok();
+        var useridstring = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        // return Ok(useridstring);
+        if (string.IsNullOrEmpty(useridstring)) return Unauthorized();
+        int userid = int.Parse(useridstring);
+        wishlistDTO.UserId = userid;
+        var result = await _wishlistservice.CreateService(wishlistDTO);
+        return Ok(result);
     }
 
     [HttpPut("update")]
     public async Task<ActionResult> UpdateWishlist(WishlistDTO wishlistDTO)
     {
-        _wishlistservice.UpdateService(wishlistDTO);
+        await _wishlistservice.UpdateService(wishlistDTO);
         return Ok();
     }
 
-    [HttpDelete("delete")]
-    public async Task<ActionResult> DeleteWishlist(int id)
+    [Authorize]
+    [HttpDelete("delete/{id:int}")]
+    public async Task<ActionResult> DeleteWishlist([FromRoute] int id)
     {
-        _wishlistservice.DeleteService(id);
+        var useridstring = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        // return Ok(useridstring);
+        if (string.IsNullOrEmpty(useridstring)) return Unauthorized();
+        int userid = int.Parse(useridstring);
+        await _wishlistservice.DeleteService(id);
+        return Ok();
+    }
+    [Authorize]
+    [HttpDelete("delete/product")]
+    public async Task<ActionResult> DeleteProductFromWishlist([FromBody] DeleteProductDto deleteProductDto)
+    {
+        var useridstring = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        // return Ok(useridstring);
+        if (string.IsNullOrEmpty(useridstring)) return Unauthorized();
+        int userid = int.Parse(useridstring);
+        await _wishlistservice.DeleteProduct(deleteProductDto.Id, deleteProductDto.WishlistName);
         return Ok();
     }
 }
